+++ resources/js/beatmap-discussions/review-post.tsx
@@ -3,11 +3,9 @@
 
 import { PersistedBeatmapDiscussionReview } from 'interfaces/beatmap-discussion-review';
 import * as React from 'react';
-import ReactMarkdown, { uriTransformer } from 'react-markdown';
-import { ReactMarkdownProps } from 'react-markdown/lib/complex-types';
-import { propsFromHref } from 'utils/beatmapset-discussion-helper';
+import ReactMarkdown from 'react-markdown';
 import { uuid } from 'utils/seq';
-import { openBeatmapEditor } from 'utils/url';
+import { linkRenderer, paragraphRenderer, transformLinkUri } from './renderers';
 import { ReviewPostEmbed } from './review-post-embed';
 
 interface Props {
@@ -28,10 +26,10 @@ export class ReviewPost extends React.Component<Props> {
       <ReactMarkdown
         key={uuid()}
         components={{
-          a: this.linkRenderer,
-          p: this.paragraphRenderer,
+          a: linkRenderer,
+          p: (node) => <div className='beatmap-discussion-review-post__block'>{paragraphRenderer(node)}</div>,
         }}
-        transformLinkUri={this.transformLinkUri}
+        transformLinkUri={transformLinkUri}
         unwrapDisallowed
       >
         {source}
@@ -69,58 +67,4 @@ export class ReviewPost extends React.Component<Props> {
       </div>
     );
   }
-
-  private linkRenderer = (astProps: ReactMarkdownProps & React.DetailedHTMLProps<React.AnchorHTMLAttributes<HTMLAnchorElement>, HTMLAnchorElement>) => {
-    const props = propsFromHref(astProps.href ?? '');
-
-    return <a href={astProps.href} {...props}>{astProps.children}</a>;
-  };
-
-  private paragraphDecorator = (reactNode: React.ReactNode) => {
-    if (typeof reactNode === 'string') {
-      const matches = [...reactNode.matchAll(/\b(((\d{2,}):([0-5]\d)[:.](\d{3}))(\s\((?:\d+[,|])*\d+\))?)/g)];
-
-      if (matches.length > 0) {
-        let cursor = 0;
-        const nodes: React.ReactNode[] = [];
-
-        for (const match of matches) {
-          // insert any preceding text
-          const index = match.index ?? 0;
-          const textFragment = reactNode.substring(cursor, index);
-          nodes.push(textFragment);
-
-          // decorate the timestamp as a link
-          const timestamp = match[0];
-          nodes.push((
-            <a key={`timestamp-${index}`} className='beatmap-discussion-timestamp-decoration' href={openBeatmapEditor(timestamp)}>{timestamp}</a>
-          ));
-
-          cursor = index + match[0].length;
-        }
-
-        // add any remaining text
-        nodes.push(reactNode.substring(cursor, reactNode.length));
-
-        return nodes;
-      }
-    }
-
-    return reactNode;
-  };
-
-  private paragraphRenderer = (astProps: ReactMarkdownProps & React.DetailedHTMLProps<React.HTMLAttributes<HTMLParagraphElement>, HTMLParagraphElement>) => (
-    <div className='beatmap-discussion-review-post__block'>
-      <div className='beatmapset-discussion-message'>{astProps.children.map(this.paragraphDecorator)}</div>
-    </div>
-  );
-
-  private transformLinkUri = (uri: string) => {
-    if (uri.startsWith('osu://edit/')) {
-      // TODO: sanitize timestamp?
-      return uri;
-    }
-
-    return uriTransformer(uri);
-  };
 }
